Moved initialization file for motor QEI and IC in a new file called motors/init.c
