+++ includes/system/system.h
@@ -13,7 +13,7 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details
-*/
+ */
 
 #ifndef SYSTEM_H
 #define	SYSTEM_H
@@ -22,69 +22,70 @@
 extern "C" {
 #endif
 
-    #include "packet/packet.h"
+#include "packet/packet.h"
 
-/******************************************************************************/
-/* System Level #define Macros                                                */
-/******************************************************************************/
-    
+    /******************************************************************************/
+    /* System Level #define Macros                                                */
+    /******************************************************************************/
+
     /* Name interrupt */
-    #define VEL_PID_PRIORITY IPC0bits.OC1IP
-    #define RX_PARSER_PRIORITY IPC1bits.OC2IP
-    #define MEASURE_PRIORITY IPC6bits.OC3IP
-    #define DEAD_RECK_PRIORITY IPC15bits.RTCIP
+#define VEL_PID_PRIORITY IPC0bits.OC1IP
+#define RX_PARSER_PRIORITY IPC1bits.OC2IP
+#define MEASURE_PRIORITY IPC6bits.OC3IP
+#define DEAD_RECK_PRIORITY IPC15bits.RTCIP
 
     /* Interrupt enable */
-    #define VEL_PID_ENABLE IEC0bits.OC1IE
-    #define RX_PARSER_ENABLE IEC0bits.OC2IE
-    #define MEASURE_ENABLE IEC1bits.OC3IE
-    #define DEAD_RECK_ENABLE IEC3bits.RTCIE
+#define VEL_PID_ENABLE IEC0bits.OC1IE
+#define RX_PARSER_ENABLE IEC0bits.OC2IE
+#define MEASURE_ENABLE IEC1bits.OC3IE
+#define DEAD_RECK_ENABLE IEC3bits.RTCIE
 
     /* Interrupt priority */
     /* Max priority 7 - Min priority 1 */
-    #define PWM_TIMER_LEVEL 7
-    #define INPUT_CAPTURE_LEVEL 6
-    #define ADC_DMA_LEVEL 6
-    #define UART_RX_LEVEL 6
-    #define SYS_TIMER_LEVEL 5
-    #define MEASURE_LEVEL 4
-    #define CURR_PID_LEVEL 4
-    #define VEL_PID_LEVEL 3
-    #define DEAD_RECK_LEVEL 2
-    #define UART_TX_LEVEL 1
-    #define RX_PARSER_LEVEL 1
+#define PWM_TIMER_LEVEL 7
+#define INPUT_CAPTURE_LEVEL 6
+#define ADC_DMA_LEVEL 6
+#define UART_RX_LEVEL 6
+#define SYS_TIMER_LEVEL 5
+#define MEASURE_LEVEL 4
+#define CURR_PID_LEVEL 4
+#define VEL_PID_LEVEL 3
+#define DEAD_RECK_LEVEL 2
+#define UART_TX_LEVEL 1
+#define RX_PARSER_LEVEL 1
 
     /* Microcontroller MIPs (FCY) */
-    #define SYS_FREQ        80000000
-    #define FCY             SYS_FREQ/2
+#define SYS_FREQ        80000000
+#define FCY             SYS_FREQ/2
 
-    #define FRTMR1 1000             // Timer1 - Value in herz [Hz]
-    #define TCTMR1 1/FRTMR1         // Timer1 - Value in seconds [s]
-    #define TMR1_VALUE FCY/FRTMR1   // Timer1 - Value in CLK
-    #define FRTMR2 SYS_FREQ
-    #define TMR2_VALUE 0xFFFF       // Timer2 - Value for overflow
+#define FRTMR1 1000             // Timer1 - Value in herz [Hz]
+#define TCTMR1 1/FRTMR1         // Timer1 - Value in seconds [s]
+#define TMR1_VALUE FCY/FRTMR1   // Timer1 - Value in CLK
+#define FRTMR2 SYS_FREQ
+#define TMR2_VALUE 0xFFFF       // Timer2 - Value for overflow
 
-//    //Blink LED
-//    /**
-//     * BL = 0.5 = 1/2
-//     * BLINKSW = BL/0.001 = (1/2)/10^-3 = 10^3/2 = 1000/2 = 500
-//     */
-//    #define BLINK_LED 0.5 //Value in seconds [s]
-//    #define BLINKSW (int)(BLINK_LED/TCTMR1)
+    //    //Blink LED
+    //    /**
+    //     * BL = 0.5 = 1/2
+    //     * BLINKSW = BL/0.001 = (1/2)/10^-3 = 10^3/2 = 1000/2 = 500
+    //     */
+    //    #define BLINK_LED 0.5 //Value in seconds [s]
+    //    #define BLINKSW (int)(BLINK_LED/TCTMR1)
 
     //UART
-    #define BAUDRATE 115200
+#define BAUDRATE 115200
     //#define BAUDRATE 57600
-    #define BRGVAL   ((FCY/BAUDRATE)/16)-1
+#define BRGVAL   ((FCY/BAUDRATE)/16)-1
 
     // Current ADC buffer dimension
-    #define ADC_CHANNELS 2
-    #define ADC_BUFF 64
-    #define TOT_ADC_BUFF ADC_CHANNELS * ADC_BUFF
+#define ADC_CHANNELS 2
+#define ADC_BUFF 64
+#define TOT_ADC_BUFF ADC_CHANNELS * ADC_BUFF
 
 #define NUM_PROCESS_DEFAULT 2
-    #define PROCESS_IDLE 0
-    #define PROCESS_PARSE 1
+#define PROCESS_IDLE 0
+#define PROCESS_PARSE 1
+
     typedef struct process {
         char name[MAX_BUFF_TASK_NAME];
         uint8_t time;
@@ -92,13 +93,13 @@ extern "C" {
         uint8_t frequency;
     } process_t;
 
-/******************************************************************************/
-/* System Function Prototypes                                                 */
-/******************************************************************************/
+    /******************************************************************************/
+    /* System Function Prototypes                                                 */
+    /******************************************************************************/
 
-/* Custom oscillator configuration funtions, reset source evaluation
-functions, and other non-peripheral microcontroller initialization functions
-go here. */
+    /* Custom oscillator configuration funtions, reset source evaluation
+    functions, and other non-peripheral microcontroller initialization functions
+    go here. */
 
     /**
      * Initialization name process and set standard priority for all procesees
@@ -120,7 +121,7 @@ go here. */
      * @return ACK value for correct update priority
      */
     unsigned char update_frequency(void);
-    
+
     void set_process(uint8_t command, system_task_t process_state);
     /**
      * From name recived, return a process required.
@@ -140,30 +141,6 @@ go here. */
     /** Handles clock switching/osc initialization
      */
     void ConfigureOscillator(void);
-
-    /** Initialization PWM for drive engine left & right
-     */
-    void InitPWM(void);
-
-    /** Initialization QEI - Motor left
-     */
-    void InitQEI1(void);
-
-    /** Initialization QEI - Motor right
-     */
-    void InitQEI2(void);
-
-    /** Initialization Input Capture - Motor left
-     */
-    void InitIC1(void);
-
-    /** Initialization Input Capture - Motor right
-     */
-    void InitIC2(void);
-
-    /** Safely switch to the new Input Capture prescaler
-     */
-    void SwitchIcPrescaler(int mode, int motIdx);
 
     /** Initialization Timer 1 - Timer system
      */
