+++ src/communication/parsing_other_messages.c
@@ -38,7 +38,7 @@
 #include "communication/serial.h"
 
 #include "control/high_level_control.h"
-#include "control/motors.h"
+#include "control/motors/motors.h"
 
 #include "system/user.h"
 #include "system/system.h"
